+++ src/main/java/com/contentful/java/cma/model/CMASystem.java
@@ -1,5 +1,7 @@
 package com.contentful.java.cma.model;
 
+import com.google.gson.annotations.SerializedName;
+
 import java.util.HashMap;
 
 /**
@@ -25,6 +27,9 @@ public class CMASystem {
   CMALink updatedBy;
   Integer version;
   Integer archivedVersion;
+
+  @SerializedName("status")
+  CMALink environmentStatus;
 
   /**
    * @return the content type if this resource can have one.
@@ -208,6 +213,17 @@ public class CMASystem {
   }
 
   /**
+   * Return a link to the state of this environment
+   *
+   * This field is only active on an environment.
+   * @return a link to the status.
+   * @see CMAEnvironment#getStatus()
+   */
+  public CMALink getEnvironmentalStatus() {
+    return environmentStatus;
+  }
+
+  /**
    * @return a human readable string, representing the object.
    */
   @Override public String toString() {
@@ -229,6 +245,7 @@ public class CMASystem {
     map.put("updatedAt", getUpdatedAt());
     map.put("updatedBy", getUpdatedBy());
     map.put("version", getVersion());
+    map.put("status", getEnvironmentalStatus());
 
     final StringBuilder builder = new StringBuilder("CMASystem { ");
     String separator = "";